+++ spec/rrj/rabbit/publish/base_publisher_spec.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe RubyRabbitmqJanus::Rabbit::Publisher, type: :rabbit,
+                                               name: :base_publisher do
+  let(:publish) { RubyRabbitmqJanus::Rabbit::Publisher::BasePublisher.new }
+
+  describe '#new' do
+    it { expect(publish).to have_attributes(response: nil) }
+  end
+end