+++ spec/rrj/rabbit/publish/listener_spec.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe RubyRabbitmqJanus::Rabbit::Publisher, type: :rabbit,
+                                               name: :listener do
+  let(:publish) do
+    rabbit = RubyRabbitmqJanus::Rabbit::Connect.new.rabbit
+    RubyRabbitmqJanus::Rabbit::Publisher::Listener.new(rabbit)
+  end
+
+  describe '#new' do
+    it { expect(publish).to have_attributes(response: nil) }
+  end
+end